Convert byte encoded headers to strings
For some reason aimbat sets header values to bytes instead of
strings in the pickle files. This means rstrip() doesn't work
and the values are too long. Should really be fixed in aimbat,
but since both aimbat and this branch of pysmo are going to be
obsolete it is easier to fix here.